+++ src/sandboxcli/domain/connection/value_objects/__init__.py
@@ -2,10 +2,12 @@
 from .auth_credential import AuthCredential
 from .connection_config import ConnectionConfig
 from .connection_status import ConnectionStatus, ConnectionStatusEnum
+from .health_check_result import HealthCheckResult
 
 __all__ = [
     "AuthCredential",
     "ConnectionConfig",
     "ConnectionStatus",
     "ConnectionStatusEnum",
+    "HealthCheckResult",
 ]